app/requirements/development.txt
-r base.txt
django-extensions
werkzeug
django-debug-toolbar
django-nose
